Simplify restart connect
